requirements.txt
flask==2.1.2
flask-limiter==2.0.0
flask-swagger-ui==3.36.0
transformers==4.10.0
torch==1.9.0
flask-caching==1.10.1
pyjwt==2.1.0
